split loss modules
